feat(manager-react-components): delete deprecated components
ref: #MANAGER-19193
