+++ packages/manager-react-components/src/components/index.ts
@@ -41,5 +41,5 @@ export * from './region/region.component';
 export * from './order';
 
 export * from './badge';
-export * from './Modal';
+export * from './modal';
 export * from './tags-list';
