+++ packages/manager-react-components/src/components/service-state-badge/ServiceStateBadge.props.ts
@@ -1,6 +1,13 @@
 import { ComponentProps } from 'react';
-import { ResourceStatus } from '../../hooks/services/services.type';
 import { Badge } from '../badge';
+
+export type ResourceStatus =
+  | 'active'
+  | 'deleted'
+  | 'suspended'
+  | 'toActivate'
+  | 'toDelete'
+  | 'toSuspend';
 
 export type ServiceStateBadgeProps = Omit<
   ComponentProps<typeof Badge>,
