+++ packages/manager-react-components/src/components/service-state-badge/__tests__/ServiceStateBadge.spec.util.ts
@@ -1,4 +1,4 @@
-import { ResourceStatus } from '../../../hooks';
+import { ResourceStatus } from '../ServiceStateBadge.props';
 
 export const SERVICE_STATES = [
   {
