+++ packages/manager-react-components/src/components/templates/index.ts
@@ -5,6 +5,4 @@ export * from './error-boundary/error-boundary.component';
 export * from './onboarding/onboarding.component';
 export * from './layout/layout.component';
 export * from './delete-modal/delete-modal.component';
-export * from './delete-modal/delete-service-modal.component';
 export * from './update-name-modal/update-name-modal.component';
-export * from './update-name-modal/update-iam-name-modal.component';
