--- packages/manager-react-components/src/hooks/feature-availability/index.ts
@@ -1,2 +1,0 @@
-export * from './useFeatureAvailability';
-export * from './mocks/feature-availability.mock';
