+++ packages/manager-react-components/src/hooks/index.ts
@@ -1,16 +1,9 @@
-import { TRegion as Region } from './useProjectRegions';
-
 export * from './breadcrumb/useBreadcrumb';
 export * from './useCatalogPrice';
 export * from './useMe';
-export * from './useProjectRegions';
-export * from './useProjectUrl';
-export * from './feature-availability';
 export * from './datagrid/useIcebergV2';
 export * from './datagrid/useIcebergV6';
 export * from './datagrid/useResourcesV6';
-export * from './services';
-export * from './tasks';
 export * from './date';
 export * from './iam';
 export * from './bytes/useBytes';
@@ -19,5 +12,3 @@ export {
   useTranslatedMicroRegions,
   isLocalZone,
 } from './region/useTranslatedMicroRegions';
-
-export type TRegion = Region;
