--- packages/manager-react-components/src/hooks/services/api/index.ts
@@ -1,3 +1,0 @@
-export * from './get';
-export * from './put';
-export * from './post';
