--- packages/manager-react-components/src/hooks/services/api/put.ts
@@ -1,19 +1,0 @@
-import { apiClient } from '@ovh-ux/manager-core-api';
-
-export type UpdateServiceNameParams = {
-  /** Service id */
-  serviceId: number;
-  /** Service new display name */
-  displayName: string;
-};
-
-/**
- * Update a service's display name
- */
-export const updateServiceName = async ({
-  serviceId,
-  displayName,
-}: UpdateServiceNameParams) =>
-  apiClient.v6.put(`/services/${serviceId}`, {
-    displayName,
-  });
