--- packages/manager-react-components/src/hooks/services/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './useDeleteService';
-export * from './useUpdateServiceName';
-export * from './useServiceDetails';
