--- packages/manager-react-components/src/hooks/services/index.ts
@@ -1,4 +1,0 @@
-export * from './api';
-export * from './hooks';
-export * from './mocks';
-export * from './services.type';
