--- packages/manager-react-components/src/hooks/services/mocks/index.ts
@@ -1,2 +1,0 @@
-export * from './services.handler';
-export * from './services.mock';
