--- packages/manager-react-components/src/hooks/tasks/index.ts
@@ -1,1 +1,0 @@
-export * from './hooks/useTask';
