+++ packages/manager-react-components/src/lib.ts
@@ -1,7 +1,6 @@
 import './lib.scss';
 
 export * from './hooks';
-export * from './hooks/feature-availability/useFeatureAvailability';
 export * from './components';
 export * from './enumTypes';
 export * from './utils/click-utils';
